requirements.txt
feedparser==6.0.11
nbformat>=5.10.0
nbconvert>=7.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
html2text>=2024.2.26
requests>=2.32.0
python-dateutil>=2.9.0
